fixing diff, renaming
